--- src/taskname.js
@@ -1,13 +1,0 @@
-const tasks = [
-  {
-    description: 'wash the dishes',
-    completed: true,
-    index: 1,
-  },
-  {
-    description: 'complete to do list',
-    completed: true,
-    index: 2,
-  },
-];
-export default tasks;
